Split stores into separate files (#139)
* include index.js to maintain access via $lib/data/stores
* add showDesktopSidebar store
