+++ src/lib/data/stores/storage.js
@@ -1,0 +1,5 @@
+export const setDefaultStorage = (name, value) => {
+    if (!localStorage.getItem(name) && value) {
+        localStorage.setItem(name, value)
+    }
+}
